+++ Helpers/OptimizationOptions.cs
@@ -466,9 +466,6 @@ internal partial class OptimizationOptions
                     case "EnableRecommendedSectionStartMenu":
                         OptimizeSystemHelper.EnableRecommendedSectionStartMenu();
                         break;
-                    case "EnableUnnecessaryServices":
-                        OptimizeSystemHelper.EnableUnnecessaryServices();
-                        break;
                     case "DisableLegacyBootMenu":
                         OptimizeSystemHelper.DisableLegacyBootMenu();
                         break;
@@ -796,19 +793,6 @@ internal partial class OptimizationOptions
                     {
                         OptimizeSystemHelper.DisableLegacyBootMenu();
                         RemoveRevertAction("DisableLegacyBootMenu");
-                    }
-                    break;
-
-                case "UnnecessaryServices":
-                    if (toggleSwitch.IsOn)
-                    {
-                        OptimizeSystemHelper.DisableUnnecessaryServices();
-                        SaveRevertAction("EnableUnnecessaryServices");
-                    }
-                    else
-                    {
-                        OptimizeSystemHelper.EnableUnnecessaryServices();
-                        RemoveRevertAction("EnableUnnecessaryServices");
                     }
                     break;